Initial auth test for post verify
