+++ tests/aswwu/behaviors/auth/subtests.py
@@ -1,21 +1,24 @@
-import requests
+import tests.aswwu.behaviors.auth.requests as auth_requests
+import tests.utils as utils
+import tests.aswwu.data.paths as paths
+import json
 from settings import keys, testing
 
-
-def post_verify(wwuid, full_name, email):
-    url = testing['base_url'] + testing['port'] + 'verify'
-    post_data = {
-        'secret_key': keys["samlEndpointKey"],
-        'employee_id': wwuid,
-        'full_name': full_name,
-        'email_address': email,
-    }
-    resp = requests.post(url, post_data)
-    return resp
+def send_get_verify():
+    resp = auth_requests.get_verify()
 
 
-def get_verify():
-    url = testing['base_url'] + testing['port'] + 'verify'
-    resp = requests.get(url)
-    assert (resp.status_code == 200)
-    return resp
+def send_post_verify():
+    DEFAULT_STATUS = 'Student'
+    DEFAULT_ROLES = 'None'
+    users = utils.load_users(paths.USER_PATH)
+    for user in users:
+        resp = auth_requests.post_verify(user['wwuid'], user['full_name'], user['email'])
+        resp_text = json.loads(resp.text)
+        assert (resp.status_code == 200)
+        assert (resp_text['token'].split('|')[0] == user['wwuid'])
+        assert (resp_text['user']['username'] == user['username'])
+        assert (resp_text['user']['wwuid'] == user['wwuid'])
+        assert (resp_text['user']['roles'] == DEFAULT_ROLES)
+        assert (resp_text['user']['status'] == DEFAULT_STATUS)
+        assert (resp_text['user']['full_name'] == user['full_name'])
