+++ tests/aswwu/behaviors/auth/tests.py
@@ -1,0 +1,5 @@
+import tests.aswwu.behaviors.auth.subtests as subtests
+import tests.conftest.testing_server as testing_server
+
+def test_post_verify(testing_server):
+    subtests.send_post_verify()
